# openapi.json
{"openapi":"3.1.0","info":{"title":"Bethel Trading Technologies API","description":"Institutional Algorithmic Trading & Investor Portal","version":"1.0"},"paths":{"/accounts/":{"get":{"tags":["Trading Accounts"],"summary":"Get Accounts","operationId":"get_accounts_accounts__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/data":{"get":{"tags":["Dashboard"],"summary":"Dashboard Data","operationId":"dashboard_data_dashboard_data_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analytics/performance":{"get":{"tags":["Analytics"],"summary":"Performance","operationId":"performance_analytics_performance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analytics/equity":{"get":{"tags":["Analytics"],"summary":"Equity Curve","operationId":"equity_curve_analytics_equity_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analytics/status":{"get":{"tags":["Analytics"],"summary":"Analytics Status","operationId":"analytics_status_analytics_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/risk/status":{"get":{"tags":["Risk"],"summary":"Risk Status","operationId":"risk_status_risk_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/connect":{"post":{"tags":["MT5 Connection"],"summary":"Connect Mt5","operationId":"connect_mt5_mt5_connect_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MT5ConnectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mt5/account":{"get":{"tags":["MT5 Connection"],"summary":"Account Info","operationId":"account_info_mt5_account_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/positions":{"get":{"tags":["MT5 Connection"],"summary":"Open Positions","operationId":"open_positions_mt5_positions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/test-order":{"post":{"tags":["MT5 Connection"],"summary":"Test Order","operationId":"test_order_mt5_test_order_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/symbols":{"get":{"tags":["MT5 Connection"],"summary":"Get Symbols","operationId":"get_symbols_mt5_symbols_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/history":{"get":{"tags":["MT5 Connection"],"summary":"Trade History","operationId":"trade_history_mt5_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mt5/import-history":{"post":{"tags":["MT5 Connection"],"summary":"Import History","operationId":"import_history_mt5_import_history_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/performance/equity-history":{"get":{"tags":["Performance History"],"summary":"Equity History","operationId":"equity_history_performance_equity_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/performance/analytics":{"get":{"tags":["Performance History"],"summary":"Analytics","operationId":"analytics_performance_analytics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/performance/daily":{"get":{"tags":["Performance History"],"summary":"Daily Performance","operationId":"daily_performance_performance_daily_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/performance/monthly":{"get":{"tags":["Performance History"],"summary":"Monthly Performance","operationId":"monthly_performance_performance_monthly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/performance/trades":{"get":{"tags":["Performance History"],"summary":"Trade Performance","operationId":"trade_performance_performance_trades_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investor/auth/login":{"post":{"tags":["Investor Authentication"],"summary":"Login","operationId":"login_investor_auth_login_post","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/portfolio/create":{"post":{"tags":["Portfolio"],"summary":"Create Portfolio","operationId":"create_portfolio_portfolio_create_post","parameters":[{"name":"investor_id","in":"query","required":true,"schema":{"type":"integer","title":"Investor Id"}},{"name":"starting_capital","in":"query","required":true,"schema":{"type":"number","title":"Starting Capital"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/portfolio/{investor_id}":{"get":{"tags":["Portfolio"],"summary":"Get Portfolios","operationId":"get_portfolios_portfolio__investor_id__get","parameters":[{"name":"investor_id","in":"path","required":true,"schema":{"type":"integer","title":"Investor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/portfolio/{portfolio_id}/update":{"put":{"tags":["Portfolio"],"summary":"Update Portfolio","operationId":"update_portfolio_portfolio__portfolio_id__update_put","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"integer","title":"Portfolio Id"}},{"name":"current_value","in":"query","required":true,"schema":{"type":"number","title":"Current Value"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mt5-accounts/assign":{"post":{"tags":["MT5 Accounts"],"summary":"Assign Account","operationId":"assign_account_mt5_accounts_assign_post","parameters":[{"name":"investor_id","in":"query","required":true,"schema":{"type":"integer","title":"Investor Id"}},{"name":"portfolio_id","in":"query","required":true,"schema":{"type":"integer","title":"Portfolio Id"}},{"name":"mt5_login","in":"query","required":true,"schema":{"type":"string","title":"Mt5 Login"}},{"name":"broker","in":"query","required":true,"schema":{"type":"string","title":"Broker"}},{"name":"server","in":"query","required":true,"schema":{"type":"string","title":"Server"}},{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"USD","title":"Currency"}},{"name":"initial_balance","in":"query","required":false,"schema":{"type":"number","default":0,"title":"Initial Balance"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mt5-accounts/{investor_id}":{"get":{"tags":["MT5 Accounts"],"summary":"Get Accounts","operationId":"get_accounts_mt5_accounts__investor_id__get","parameters":[{"name":"investor_id","in":"path","required":true,"schema":{"type":"integer","title":"Investor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investor/":{"get":{"tags":["Investor"],"summary":"Investor Page","operationId":"investor_page_investor__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/investor/api/performance":{"get":{"tags":["Investor"],"summary":"Investor Performance","operationId":"investor_performance_investor_api_performance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/investor/api/equity":{"get":{"tags":["Investor"],"summary":"Investor Equity","operationId":"investor_equity_investor_api_equity_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/investor/api/risk":{"get":{"tags":["Investor"],"summary":"Investor Risk","operationId":"investor_risk_investor_api_risk_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/investor/api/mt5":{"get":{"tags":["Investor"],"summary":"Investor Mt5","operationId":"investor_mt5_investor_api_mt5_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/investor/api/dashboard/{investor_id}":{"get":{"tags":["Investor Dashboard"],"summary":"Investor Dashboard","operationId":"investor_dashboard_investor_api_dashboard__investor_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"investor_id","in":"path","required":true,"schema":{"type":"integer","title":"Investor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"get":{"summary":"Login Page","operationId":"login_page_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Dashboard Home","operationId":"dashboard_home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MT5ConnectionRequest":{"properties":{"account_number":{"type":"string","title":"Account Number"},"server":{"type":"string","title":"Server"},"password":{"type":"string","title":"Password"}},"type":"object","required":["account_number","server","password"],"title":"MT5ConnectionRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}